feat(#601): add db helper
